Move some things around to make easier to use 1 connection to the Audio DB
